+++ packages/block-editor/src/experiments.js
@@ -1,19 +1,10 @@
 /**
- * WordPress dependencies
- */
-import { __dangerousOptInToUnstableAPIsOnlyForCoreModules } from '@wordpress/experiments';
-
-/**
  * Internal dependencies
  */
 import * as globalStyles from './components/global-styles';
 import { ExperimentalBlockEditorProvider } from './components/provider';
-
-export const { lock, unlock } =
-	__dangerousOptInToUnstableAPIsOnlyForCoreModules(
-		'I know using unstable features means my plugin or theme will inevitably break on the next WordPress release.',
-		'@wordpress/block-editor'
-	);
+import { lock } from './lock-unlock';
+import OffCanvasEditor from './components/off-canvas-editor';
 
 /**
  * Experimental @wordpress/block-editor APIs.
@@ -22,4 +13,5 @@ export const experiments = {};
 lock( experiments, {
 	...globalStyles,
 	ExperimentalBlockEditorProvider,
+	OffCanvasEditor,
 } );