+++ packages/block-editor/src/hooks/custom-class-name.js
@@ -59,7 +59,7 @@ export const withInspectorControl = createHigherOrderComponent(
 				return (
 					<>
 						<BlockEdit { ...props } />
-						<InspectorControls __experimentalGroup="advanced">
+						<InspectorControls group="advanced">
 							<TextControl
 								__nextHasNoMarginBottom
 								autoComplete="off"