+++ packages/block-editor/src/store/actions.js
@@ -16,7 +16,6 @@ import { speak } from '@wordpress/a11y';
 import { __, _n, sprintf } from '@wordpress/i18n';
 import { create, insert, remove, toHTMLString } from '@wordpress/rich-text';
 import deprecated from '@wordpress/deprecated';
-import { Platform } from '@wordpress/element';
 
 /**
  * Internal dependencies
@@ -26,6 +25,7 @@ import {
 	retrieveSelectedAttribute,
 	START_OF_SELECTED_AREA,
 } from '../utils/selection';
+import { __experimentalUpdateSettings } from './private-actions';
 
 /** @typedef {import('../components/use-on-block-drop/types').WPDropOperation} WPDropOperation */
 
@@ -1265,28 +1265,6 @@ export function toggleBlockMode( clientId ) {
 }
 
 /**
- * Returns an action object used in signalling that the block interface, eg. toolbar, outline, etc. should be hidden.
- *
- * @return {Object} Action object.
- */
-export function __experimentalHideBlockInterface() {
-	return {
-		type: 'HIDE_BLOCK_INTERFACE',
-	};
-}
-
-/**
- * Returns an action object used in signalling that the block interface, eg. toolbar, outline, etc. should be shown.
- *
- * @return {Object} Action object.
- */
-export function __experimentalShowBlockInterface() {
-	return {
-		type: 'SHOW_BLOCK_INTERFACE',
-	};
-}
-
-/**
  * Returns an action object used in signalling that the user has begun to type.
  *
  * @return {Object} Action object.
@@ -1444,45 +1422,6 @@ export function updateBlockListSettings( clientId, settings ) {
  */
 export function updateSettings( settings ) {
 	return __experimentalUpdateSettings( settings, true );
-}
-
-/**
- * A list of private/experimental block editor settings that
- * should not become a part of the WordPress public API.
- * BlockEditorProvider will remove these settings from the
- * settings object it receives.
- *
- * @see https://github.com/WordPress/gutenberg/pull/46131
- */
-const privateSettings = [ '__unstableInserterMediaCategories' ];
-
-/**
- * Action that updates the block editor settings and
- * conditionally preserves the experimental ones.
- *
- * @param {Object}  settings                  Updated settings
- * @param {boolean} stripExperimentalSettings Whether to strip experimental settings.
- * @return {Object} Action object
- */
-export function __experimentalUpdateSettings(
-	settings,
-	stripExperimentalSettings = false
-) {
-	let cleanSettings = settings;
-	// There are no plugins in the mobile apps, so there is no
-	// need to strip the experimental settings:
-	if ( stripExperimentalSettings && Platform.OS === 'web' ) {
-		cleanSettings = {};
-		for ( const key in settings ) {
-			if ( ! privateSettings.includes( key ) ) {
-				cleanSettings[ key ] = settings[ key ];
-			}
-		}
-	}
-	return {
-		type: 'UPDATE_SETTINGS',
-		settings: cleanSettings,
-	};
 }
 
 /**