+++ packages/block-editor/src/store/index.js
@@ -1,18 +1,18 @@
 /**
  * WordPress dependencies
  */
-import { createReduxStore, register } from '@wordpress/data';
+import { createReduxStore, registerStore } from '@wordpress/data';
 
 /**
  * Internal dependencies
  */
 import reducer from './reducer';
 import * as selectors from './selectors';
-import * as allActions from './actions';
+import * as privateActions from './private-actions';
+import * as privateSelectors from './private-selectors';
+import * as actions from './actions';
 import { STORE_NAME } from './constants';
-import { unlock } from '../experiments';
-
-const { __experimentalUpdateSettings, ...actions } = allActions;
+import { unlock } from '../lock-unlock';
 
 /**
  * Block editor data store configuration.
@@ -35,8 +35,11 @@ export const store = createReduxStore( STORE_NAME, {
 	persist: [ 'preferences' ],
 } );
 
-register( store );
-
-unlock( store ).registerPrivateActions( {
-	__experimentalUpdateSettings,
+// We will be able to use the `register` function once we switch
+// the "preferences" persistence to use the new preferences package.
+const registeredStore = registerStore( STORE_NAME, {
+	...storeConfig,
+	persist: [ 'preferences' ],
 } );
+unlock( registeredStore ).registerPrivateActions( privateActions );
+unlock( registeredStore ).registerPrivateSelectors( privateSelectors );