+++ packages/components/src/toolbar/toolbar-button/toolbar-button-container.js
@@ -1,3 +1,5 @@
+// @ts-nocheck
+
 const ToolbarButtonContainer = ( props ) => (
 	<div className={ props.className }>{ props.children }</div>
 );